requirements.txt
absl-py==2.3.1
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
dcor==0.6
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
fdasrsf==2.6.4
filelock==3.19.1
findiff==0.12.1
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.45.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
mpmath==1.3.0
multimethod==2.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.0
numpy==2.3.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rdata==1.0.0
referencing==0.37.0
requests==2.32.5
rich==14.1.0
rpds-py==0.30.0
scikit-datasets==0.2.5
scikit-fda==0.10.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
xarray==2025.9.0
xgboost==3.0.5
yarg==0.1.9
